Move Debugger to Mixin to make future Debugger Extension simpler
