+++ ensime_shared/ensime.py
@@ -7,6 +7,7 @@ import webbrowser
 from ensime_shared.error import Error
 from ensime_shared.util import catch, module_exists, Util
 from ensime_shared.launcher import EnsimeLauncher
+from ensime_shared.debugger import DebuggerClient
 from ensime_shared.config import gconfig, feedback
 
 from threading import Thread
@@ -67,7 +68,7 @@ commands = {
 }
 
 
-class EnsimeClient(object):
+class EnsimeClient(DebuggerClient, object):
     """Represents an Ensime client per ensime configuration path."""
 
     def __init__(self, vim, launcher, config_path):
@@ -677,22 +678,6 @@ class EnsimeClient(object):
         else:
             self.log(feedback["unhandled_response"].format(payload))
 
-    def handle_debug_output(self, call_id, payload):
-        """Handle responses `DebugOutputEvent`."""
-        self.raw_message(payload["body"].encode("ascii", "ignore"))
-
-    def handle_debug_break(self, call_id, payload):
-        """Handle responses `DebugBreakEvent`."""
-        self.raw_message(feedback["notify_break"])
-        self.debug_thread_id = payload["threadId"]
-
-    def handle_debug_backtrace(self, call_id, payload):
-        """Handle responses `DebugBacktrace`."""
-        frames = payload["frames"]
-        self.vim.command(":split backtrace.json")
-        to_json = json.dumps(frames, indent=2).split("\n")
-        self.vim.current.buffer[:] = to_json
-
     def complete(self, row, col):
         self.log("complete: in")
         pos = self.get_position(row, col)
@@ -788,45 +773,6 @@ class EnsimeClient(object):
                "typehint": "ImportSuggestionsReq",
                "file": self.path()}
         self.send_request(req)
-
-    def set_break(self, args, range=None):
-        self.log("set_break: in")
-        req = {"line": self.cursor()[0],
-               "maxResults": 10,
-               "typehint": "DebugSetBreakReq",
-               "file": self.path()}
-        self.send_request(req)
-
-    def clear_breaks(self, args, range=None):
-        self.log("clear_breaks: in")
-        self.send_request({"typehint": "DebugClearAllBreakReq"})
-
-    def debug_start(self, args, range=None):
-        self.log("debug_start: in")
-        if len(args) > 1:
-            self.send_request({
-                "typehint": "DebugAttachReq",
-                "hostname": args[0],
-                "port" :    args[1]})
-        else:
-            self.send_request({
-                "typehint": "DebugAttachReq",
-                "hostname": "localhost",
-                "port" :    "5005"})
-
-
-    def debug_continue(self, args, range=None):
-        self.log("debug_continue: in")
-        self.send_request({
-            "typehint": "DebugContinueReq",
-            "threadId": self.debug_thread_id})
-
-    def backtrace(self, args, range=None):
-        self.log("backtrace: in")
-        self.send_request({
-            "typehint": "DebugBacktraceReq",
-            "threadId": self.debug_thread_id,
-            "index": 0, "count": 100})
 
     def inspect_type(self, args, range=None):
         self.log("inspect_type: in")
